Creates file for custom HTTP hook
Instantiated a folder for custom hooks and created a JS file called
UseHttp for a future custom hook that will be a reusable fetch request
that takes two arguments: a URL and an array of dependencies.

This will allow me to call the custom hook in place of writing out
numerous fetch calls.